+++ src/SolrExpress.Core/Query/Parameter/IOffsetParameter.cs
@@ -10,11 +10,11 @@
         /// </summary>
         /// <param name="value">Value of limit</param>
         /// <returns></returns>
-        IOffsetParameter Configure(int value);
+        IOffsetParameter Configure(long value);
 
         /// <summary>
         /// Value of limit
         /// </summary>
-        int Value { get; }
+        long Value { get; }
     }
 }
